implemented type traits to detect comparability of key types
