+++ include/fkYAML/LexicalAnalyzer.hpp
@@ -28,7 +28,7 @@
 #include "fkYAML/VersioningMacros.hpp"
 #include "fkYAML/Assert.hpp"
 #include "fkYAML/Exception.hpp"
-#include "fkYAML/NodeTypeTraits.hpp"
+#include "fkYAML/TypeTraits.hpp"
 
 /**
  * @namespace fkyaml
