+++ include/fkYAML/OrderedMap.hpp
@@ -25,6 +25,7 @@
 
 #include "fkYAML/VersioningMacros.hpp"
 #include "fkYAML/Exception.hpp"
+#include "fkYAML/TypeTraits.hpp"
 
 /**
  * @namespace fkyaml
