+++ tool/clang_tidy/ClangTidyHelper.cpp
@@ -21,9 +21,9 @@
 #include "fkYAML/LexicalAnalyzer.hpp"
 #include "fkYAML/Node.hpp"
 #include "fkYAML/NodeType.hpp"
-#include "fkYAML/NodeTypeTraits.hpp"
 #include "fkYAML/OrderedMap.hpp"
 #include "fkYAML/Serializer.hpp"
+#include "fkYAML/TypeTraits.hpp"
 #include "fkYAML/VersioningMacros.hpp"
 #include "fkYAML/YAMLVersionType.hpp"
